% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copCAR.R
\name{negbinomial}
\alias{negbinomial}
\title{Family function for negative binomial GLMs.}
\usage{
negbinomial(theta = stop("'theta' must be specified."), link = "log")
}
\arguments{
\item{theta}{the dispersion parameter (must be positive).}

\item{link}{the link function, as a character string, name, or one-element character vector, specifying one of \code{log}, \code{sqrt}, or \code{identity}, or an object of class \dQuote{\code{\link[=family]{link-glm}}}}
}
\value{
An object of class \dQuote{\code{family}}, a list of functions and expressions needed to fit a negative binomial GLM.
}
\description{
Provides the information required to apply copCAR with negative binomial marginal distributions.
}
