% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copCAR.R
\name{residuals.copCAR}
\alias{residuals.copCAR}
\title{Extract model residuals.}
\usage{
\method{residuals}{copCAR}(object, type = c("deviance", "pearson",
  "response"), ...)
}
\arguments{
\item{object}{an object of class \code{copCAR}, typically the result of a call to \code{\link{copCAR}}.}

\item{type}{the type of residuals that should be returned. The alternatives are \dQuote{\code{deviance}} (default), \dQuote{\code{pearson}}, and \dQuote{\code{response}}.}

\item{\dots}{additional arguments.}
}
\value{
A vector of residuals.
}
\description{
Extract model residuals.
}
\seealso{
\code{\link{copCAR}}, \code{\link{residuals.glm}}
}
