% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copCAR.R
\name{adjacency.matrix}
\alias{adjacency.matrix}
\title{Return an adjacency matrix for a square lattice.}
\usage{
adjacency.matrix(m, n = NULL)
}
\arguments{
\item{m}{the number of rows in the lattice.}

\item{n}{the number of columns in the lattice. Defaults to \code{NULL}. If missing, the lattice is assumed to be \code{m} by \code{m}.}
}
\value{
A matrix \eqn{A} of 0s and 1s, where \eqn{A_{ij}} is equal to 1 iff vertices \eqn{i} and \eqn{j} are adjacent.
}
\description{
Return an adjacency matrix for a square lattice.
}
\details{
This function builds the adjacency matrix for the \code{m} by \code{n} square lattice.
}
